net/wireless/bcmdhd: use per DHD-instance notifier
In wl_cfg80211_netdev_notifier_call() we can not assume that net_device
passed in actually belongs to Broadcom driver; it can be any other
network device in the system. So we can not assume that we can get valid
cfg pointer from it.

To avoid crashing let's install individual notifiers per-cfg instances
and get cfg from the notifier itself.

Fixes reboots when inserting wifi adapters.

Signed-off-by: Re4son <3520622-re4son@users.noreply.gitlab.com>